Address review comments
- Rename _backend.to_numpy → as_numpy to avoid clash with conftest.to_numpy;
  use array_api_compat.{is_numpy_array,is_torch_array} for explicit dispatch
  and raise TypeError on unsupported inputs. Strengthen with resolve_conj/
  resolve_neg for torch lazy bits (no array-API equivalent).
- Move check_like and check_matrix_pair from _backend.py to utils.utils.
  check_like(None) now returns the array_api_compat.numpy namespace for
  consistency. check_matrix_pair narrows except to (ValueError, RuntimeError).
- mean_alm: replace zeros_like + [...] = with xp.asarray(copy=True).
- _get_eigenvals: take real part once at source; is_pos_def / is_pos_semi_def
  simplify accordingly.
- Drop unnecessary float() casts in _ledoit_wolf_shrinkage, ledoit_wolf, oas
  and median (preserves autograd; tests pass without cast).
- conftest.to_numpy delegates to library as_numpy; tighter scalar whitelist.
- Tests use the backend fixture instead of is_numpy_namespace(get_namespace()).
- doc/index.rst: add NumPy and PyTorch backends bullet to Key features.

+++ pyriemann/utils/median.py
@@ -80,7 +80,7 @@ def median_euclid(X, *, tol=10e-6, maxiter=50, init=None, weights=None):
         if n_zeros > 0:
             R = xp.tensordot(w, X[~is_zero] - M, axes=([0], [0]))  # Eq(2.7)
             r = xp.linalg.matrix_norm(R, ord="fro")
-            rinv = 0 if r == 0 else float(xp.mean(weights[is_zero])) / r
+            rinv = 0 if r == 0 else xp.mean(weights[is_zero]) / r
             Mnew = max(0, 1 - rinv) * Mnew + min(1, rinv) * M  # Eq(2.6)
 
         crit = xp.linalg.matrix_norm(Mnew - M, ord="fro")
